+++ lib/skylight/worker/collector.rb
@@ -1,8 +1,11 @@
+require 'uri'
+
 module Skylight
   module Worker
     class Collector < Util::Task
+      include URI::Escape
+
       ENDPOINT     = '/report'.freeze
-      FLUSH_DELAY  = 0.5
       CONTENT_TYPE = 'content-type'.freeze
       SKYLIGHT_V1  = 'application/x-skylight-report-v1'.freeze
 
@@ -11,14 +14,17 @@ module Skylight
       attr_reader :config
 
       def initialize(config)
-        super(1000)
+        super(1000, 0.25)
 
-        @config   = config
-        @http     = Util::HTTP.new(config)
-        @size     = config[:'agent.sample']
-        @batch    = nil
-        @interval = config[:'agent.interval']
-        @buf      = ""
+        @config      = config
+        @size        = config[:'agent.sample']
+        @batch       = nil
+        @interval    = config[:'agent.interval']
+        @buf         = ""
+        @refresh_at  = 0
+        @http_auth   = Util::HTTP.new(config, :accounts)
+        @http_report = nil
+        # @http_report = Util::HTTP.new(config, :report)
 
         t { fmt "starting collector; interval=%d; size=%d", @interval, @size }
       end
@@ -26,8 +32,18 @@ module Skylight
       def handle(msg, now = Util::Clock.secs)
         @batch ||= new_batch(now)
 
+        if should_refresh_token?(now)
+          refresh_report_token(now)
+        end
+
         if @batch.should_flush?(now)
-          flush(@batch)
+          if has_report_token?(now)
+            flush(@batch)
+          else
+            warn "do not have valid session token -- dropping"
+            return
+          end
+
           @batch = new_batch(now)
         end
 
@@ -47,7 +63,17 @@ module Skylight
 
       def finish
         t { fmt "collector finishing up" }
-        flush(@batch) if @batch
+
+        now = Util::Clock.secs
+
+        if should_refresh_token?(now)
+          refresh_report_token(now)
+        end
+
+        if @batch && has_report_token?(now)
+          flush(@batch)
+        end
+
         @batch = nil
       end
 
@@ -57,7 +83,42 @@ module Skylight
         debug "flushing batch; size=%d", batch.sample.count
 
         @buf.clear
-        @http.post(ENDPOINT, batch.encode(@buf), CONTENT_TYPE => SKYLIGHT_V1)
+        @http_report.post(ENDPOINT, batch.encode(@buf), CONTENT_TYPE => SKYLIGHT_V1)
+      end
+
+      def refresh_report_token(now)
+        res = @http_auth.get("/agent/authenticate?hostname=#{escape(config[:'hostname'])}")
+
+        unless res.success?
+          warn "could not fetch report session token; status=%s", res.status
+          return
+        end
+
+        tok = res.body['session']
+        tok = tok['token'] if tok
+
+        if tok
+          @refresh_at  = now + 600
+          @http_report = Util::HTTP.new(config, :report)
+          @http_report.authentication = tok
+        else
+          if @http_report
+            @refresh_at = now + 60
+          end
+          warn "server did not return a session token"
+        end
+      rescue Exception => e
+        error "exception; msg=%s; class=%s", e.message, e.class
+        t { e.backtrace.join("\n") }
+      end
+
+      def should_refresh_token?(now)
+        now >= @refresh_at
+      end
+
+      def has_report_token?(now)
+        return unless @http_report
+        now < @refresh_at + (3600 * 3 - 660)
       end
 
       def new_batch(now)
@@ -71,12 +132,12 @@ module Skylight
       class Batch
         include Util::Logging
 
-        attr_reader :config, :from, :counts, :sample
+        attr_reader :config, :from, :counts, :sample, :flush_at
 
         def initialize(config, size, from, interval)
           @config   = config
           @from     = from
-          @flush_at = from + interval + FLUSH_DELAY
+          @flush_at = from + interval
           @sample   = Util::UniformSample.new(size)
           @counts   = Hash.new(0)
         end